+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - redesign asset select modal [#362](https://github.com/asgardex/asgardex-desktop/pull/362)
 - Updated electron package [electron](https://github.com/asgardex/asgardex-desktop/pull/360)
 - Added chain Radix [#294](https://github.com/asgardex/asgardex-desktop/pull/363)
+- Remove Swapping to synths, Target asset cannot be synth [#373](https://github.com/asgardex/asgardex-desktop/pull/364)
 
 # 1.22.4 (2024-09-6)